requirements.txt
requests
python-telegram-bot==20.7
python-dotenv
selenium
webdriver-manager
